+++ main.py
@@ -1,545 +1,4 @@
-#!/usr/bin/env python3
-"""
-GMGN Wallet Copy‑Trader v7 – Sharpe‑filtered, Kelly‑sized, Jupiter‑routed
-========================================================================
-*100 % free‑API stack* ➜ gmgn.ai + Solscan + Pyth price feed + Jupiter Ultra API + public Solana RPC.
-
-Key Profit Upgrades
--------------------
-1. **Sharpe‑ratio wallet filter** – copy only risk‑adjusted winners.
-2. **Fractional‑Kelly + ATR sizing** – adaptive stake per trade.
-3. **Jupiter routing** – best price across all Solana DEXs (0.1 bps avg edge).
-4. **Priority‑fee bidding** – adds `ComputeBudgetProgram.set_compute_unit_price()` for faster fills.
-5. **Global 20 % draw‑down stop & Telegram alert.**
-6. **Prometheus metrics** – NAV, PnL, slippage exposed at `/metrics`.
-
-Free/Low‑Cost APIs Used
------------------------
-* **gmgn.ai** – wallet stats & trade list  *(free)*.
-* **Solscan public API** – holder & meta checks  *(free up to 30 req/min)*.
-* **Pyth price HTTP** – minute candles for ATR  *(free)*.
-* **Jupiter Ultra API** – `/quote` + `/swap` JSON  *(free)*.
-* **Public RPC** (api.mainnet-beta.solana.com) + optional Jito RPC (free tier).
-
-Install & run
--------------
-```bash
-pip install gmgnai-wrapper aiohttp websockets python-dotenv prometheus-client pandas
-# .env → PRIVATE_KEY, TELEGRAM_BOT_TOKEN, TELEGRAM_CHAT_ID  (optional)
-python gmgn_wallet_bot.py mirror --seed trending --min-profit 5
-```
-
-Below is the complete, production‑ready Python 3.11 script.
-"""
-# ------------------------ imports -------------------------
-from __future__ import annotations
-
-import asyncio
-import json
-import os
-import math
-import time
-import statistics
-import logging
-import base64
-import base58
-from typing import Dict, List, Sequence
-from dataclasses import dataclass
-
-import aiohttp
-import websockets
-import pandas as pd
-from dotenv import load_dotenv
-from prometheus_client import Gauge, start_http_server
-from solana.rpc.api import Client
-from solana.transaction import Transaction
-from solana.keypair import Keypair
-
-try:
-    from gmgn import gmgn
-except ImportError as exc:
-    raise SystemExit("pip install gmgnai-wrapper") from exc
-
-# ------------------------ config --------------------------
-load_dotenv()
-PRIV_KEY = os.getenv("PRIVATE_KEY") or ""
-TELE_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN", "")
-TELE_CHAT = os.getenv("TELEGRAM_CHAT_ID", "")
-STOP_LOSS_PCT = float(os.getenv("STOP_LOSS_PCT", 8))
-TAKE_PROFIT_PCT = float(os.getenv("TAKE_PROFIT_PCT", 25))
-GLOBAL_DD_PCT = float(os.getenv("GLOBAL_DD_PCT", 20))
-MAX_KELLY_F = float(os.getenv("MAX_KELLY_F", 0.25))  # cap stake pct
-ATR_LOOKBACK_MIN = int(os.getenv("ATR_LOOKBACK_MIN", 1440))
-PRUNE_INTERVAL_H = float(os.getenv("PRUNE_INTERVAL_H", 6))
-JUPITER_URL = "https://quote-api.jup.ag"
-PYTH_HIST_URL = "https://hermes.pyth.network/api/historical_price/"
-RPC_URL = os.getenv("RPC_URL", "https://api.mainnet-beta.solana.com")
-
-
-async def retry(
-    coro_fn, *, retries: int = 3, timeout: float = 10.0, name: str = "req", notif=None
-):
-    """Retry an awaitable callable with exponential backoff."""
-    delay = 1.0
-    for attempt in range(1, retries + 1):
-        try:
-            return await asyncio.wait_for(coro_fn(), timeout=timeout)
-        except Exception as exc:  # noqa: BLE001
-            logging.getLogger(__name__).warning(
-                "%s attempt %s failed: %s", name, attempt, exc
-            )
-            if attempt == retries:
-                if notif is not None:
-                    try:
-                        await notif.send(f"{name} failed: {exc}")
-                    except Exception:  # noqa: BLE001
-                        pass
-                raise
-            await asyncio.sleep(delay)
-            delay = min(delay * 2, 30)
-
-
-# ------------------------ models --------------------------
-@dataclass(slots=True)
-class Trade:
-    ts: int
-    side: str
-    token: str
-    qty: float
-    price: float
-    pnl: float
-
-
-@dataclass(slots=True)
-class WalletMetrics:
-    address: str
-    sharpe: float
-    realised: float
-    win_rate: float
-    trades: int
-
-    def is_strong(self) -> bool:
-        return self.sharpe > 1.2 and self.win_rate * 100 >= 55 and self.realised > 0
-
-
-@dataclass(slots=True)
-class Position:
-    token: str
-    qty: float
-    entry: float
-    value: float
-    sl: float
-    tp: float
-    src: str
-
-
-# ---------------- gmgn + solscan + pyth -------------------
-class GmgnAPI:  # minimal wrapper
-    def __init__(self, notif: "Notifier | None" = None):
-        self.g = gmgn()
-        self.http = aiohttp.ClientSession()
-        self.notif = notif
-
-    async def info(self, addr, tf="30d"):
-        loop = asyncio.get_running_loop()
-        return await retry(
-            lambda: loop.run_in_executor(None, self.g.getWalletInfo, addr, tf),
-            name="gmgn.info",
-            notif=self.notif,
-        )
-
-    async def trades(self, addr, tf="30d"):
-        url = f"https://gmgn.ai/stats/wallet/tx?address={addr}&period={tf}"
-
-        async def go():
-            async with self.http.get(url) as r:
-                r.raise_for_status()
-                return (await r.json()).get("data", [])
-
-        return await retry(go, name="gmgn.trades", notif=self.notif)
-
-
-class SolscanAPI:
-    BASE = "https://public-api.solscan.io"
-
-    def __init__(self, notif: "Notifier | None" = None):
-        self.http = aiohttp.ClientSession()
-        self.notif = notif
-
-    async def holders(self, mint):
-        async def go():
-            async with self.http.get(
-                f"{self.BASE}/token/holders?account={mint}&limit=50"
-            ) as r:
-                r.raise_for_status()
-                return await r.json()
-
-        return await retry(go, name="solscan.holders", notif=self.notif)
-
-    async def meta(self, mint):
-        async def go():
-            async with self.http.get(f"{self.BASE}/token/meta?account={mint}") as r:
-                r.raise_for_status()
-                return await r.json()
-
-        return await retry(go, name="solscan.meta", notif=self.notif)
-
-
-class SafetyChecker:
-    def __init__(self, sol: SolscanAPI, notif: "Notifier | None" = None):
-        self.sol = sol
-        self.notif = notif
-
-    async def _solscan_ok(self, mint: str) -> bool:
-        try:
-            holders, meta = await asyncio.gather(
-                self.sol.holders(mint), self.sol.meta(mint)
-            )
-        except Exception as exc:  # noqa: BLE001
-            logging.getLogger(__name__).warning("Solscan fail %s: %s", mint, exc)
-            if self.notif:
-                try:
-                    await self.notif.send(f"Solscan error {mint}: {exc}")
-                except Exception:  # noqa: BLE001
-                    pass
-            return False
-        if not meta or meta.get("status") != "success":
-            return False
-        return len(holders.get("data", [])) >= 5
-
-    async def rugcheck_pass(self, mint: str) -> bool:
-        base = "https://api.rugcheck.xyz/v1"
-        async with aiohttp.ClientSession() as s:
-
-            async def liq_call():
-                return await (
-                    await s.get(f"{base}/liquidity/{mint}", timeout=10)
-                ).json()
-
-            async def vote_call():
-                return await (await s.get(f"{base}/votes/{mint}", timeout=5)).json()
-
-            liq = await retry(liq_call, name="rugcheck.liq", notif=self.notif)
-            if liq["lp_owner_is_token_authority"] or liq["locked_pct"] < 70:
-                return False
-            vote = await retry(vote_call, name="rugcheck.vote", notif=self.notif)
-            return vote.get("vote") != "rug"
-
-    async def is_safe(self, mint: str) -> bool:
-        solscan_ok = await self._solscan_ok(mint)
-        return solscan_ok and await self.rugcheck_pass(mint)
-
-
-async def pyth_atr(
-    mint: str, minutes: int = ATR_LOOKBACK_MIN, notif: "Notifier | None" = None
-) -> float:
-    end = int(time.time())
-    start = end - 60 * minutes
-    url = f"{PYTH_HIST_URL}{mint}?start_time={start}&end_time={end}&interval=1"
-    async with aiohttp.ClientSession() as s:
-
-        async def go():
-            async with s.get(url) as r:
-                if r.status != 200:
-                    return None
-                data = await r.json()
-                return [p[1] for p in data.get("prices", [])][-minutes:]
-
-        prices = await retry(go, name="pyth.atr", notif=notif) or []
-    if len(prices) < 2:
-        return 0.05
-    returns = [
-        abs(prices[i] - prices[i - 1]) / prices[i - 1] for i in range(1, len(prices))
-    ]
-    return statistics.mean(returns) * math.sqrt(1440)  # daily vol approx
-
-
-async def pyth_price(mint: str, notif: "Notifier | None" = None) -> float:
-    """Return the latest Pyth price for ``mint``."""
-    end = int(time.time())
-    start = end - 120
-    url = f"{PYTH_HIST_URL}{mint}?start_time={start}&end_time={end}&interval=1"
-    async with aiohttp.ClientSession() as s:
-
-        async def go():
-            async with s.get(url) as r:
-                if r.status != 200:
-                    return None
-                data = await r.json()
-                return [p[1] for p in data.get("prices", [])]
-
-        prices = await retry(go, name="pyth.price", notif=notif) or []
-        return float(prices[-1]) if prices else 0.0
-
-
-# --------------------- analytics --------------------------
-class WalletAnalyzer:
-    def __init__(self, tf="30d", notif: "Notifier | None" = None):
-        self.tf = tf
-        self.api = GmgnAPI(notif)
-
-    async def _one(self, addr):
-        info, trades = await asyncio.gather(
-            self.api.info(addr, self.tf), self.api.trades(addr, self.tf)
-        )
-        if not info or info.get("code") != 0:
-            return None
-        pnl_by_day = {}
-        wins = 0
-        for t in trades:
-            day = time.strftime("%Y-%m-%d", time.gmtime(int(t["timestamp"]) / 1000))
-            pnl_by_day[day] = pnl_by_day.get(day, 0) + float(t["pnl"])
-            if float(t["pnl"]) > 0:
-                wins += 1
-        if len(pnl_by_day) < 2:
-            sharpe = 0
-        else:
-            series = pd.Series(list(pnl_by_day.values()))
-            sharpe = series.mean() / max(series.std(), 1e-6)
-        win_rate = wins / max(1, len(trades))
-        return WalletMetrics(
-            info["data"]["address"],
-            sharpe,
-            float(info["data"]["totalRealizedProfit"]),
-            win_rate,
-            len(trades),
-        )
-
-    async def strong(self, addrs: List[str]):
-        res = await asyncio.gather(*(self._one(a) for a in addrs))
-        return [m.address for m in res if m and m.is_strong()]
-
-
-# ------------------- execution layer ----------------------
-class JupiterExec:
-    def __init__(self, notif: "Notifier | None" = None):
-        self.http = aiohttp.ClientSession()
-        self.notif = notif
-
-    async def quote(self, mint_in, mint_out, amount):
-        url = f"{JUPITER_URL}/v6/quote?inputMint={mint_in}&outputMint={mint_out}&amount={amount}&slippageBps=50"
-
-        async def go():
-            async with self.http.get(url) as r:
-                r.raise_for_status()
-                return await r.json()
-
-        return await retry(go, name="jup.quote", notif=self.notif)
-
-    async def swap_tx(self, route):
-        async def go():
-            async with self.http.post(
-                f"{JUPITER_URL}/v6/swap",
-                json={"route": route, "userPublicKey": route["inAmountAddress"]},
-            ) as r:
-                r.raise_for_status()
-                tx = await r.json()
-                return tx["swapTransaction"]
-
-        return await retry(go, name="jup.swap", notif=self.notif)
-
-
-# priority fee helper
-def add_priority_fee(tx_bytes: bytes, lamports_per_cu: int = 1000) -> bytes:
-    # placeholder – would insert ComputeBudget ix before first ix
-    return tx_bytes
-
-
-# ---------------- position & risk -------------------------
-class PositionBook:
-    def __init__(self, init_nav: float = 100):
-        self.init = init_nav
-        self.pos: Dict[str, Position] = {}
-        self.mark: Dict[str, float] = {}
-        self.peak = init_nav
-
-    def update(self, token: str, qty: float, price: float, side: str = "buy"):
-        """Record executed trade and update mark price."""
-        self.mark[token] = price
-        pos = self.pos.get(token)
-        if side == "buy":
-            if pos:
-                new_qty = pos.qty + qty
-                pos.entry = (pos.entry * pos.qty + price * qty) / new_qty
-                pos.qty = new_qty
-                pos.value += qty * price
-            else:
-                self.pos[token] = Position(
-                    token,
-                    qty,
-                    price,
-                    qty * price,
-                    price * (1 - STOP_LOSS_PCT / 100),
-                    price * (1 + TAKE_PROFIT_PCT / 100),
-                    "copy",
-                )
-        else:
-            if pos:
-                pos.qty -= qty
-                pos.value -= qty * price
-                if pos.qty <= 0:
-                    del self.pos[token]
-
-    def nav(self):
-        free = self.init - sum(p.value for p in self.pos.values())
-        pos = sum(self.mark.get(t, p.entry) * p.qty for t, p in self.pos.items())
-        return free + pos
-
-    def update_peak(self, nav):
-        self.peak = max(self.peak, nav)
-
-    def global_dd(self, nav):
-        return 100 * (1 - nav / self.peak)
-
-
-# ---------------- telegram + prom -------------------------
-class Notifier:
-    def __init__(self):
-        self.tg_enabled = bool(TELE_TOKEN and TELE_CHAT)
-        if self.tg_enabled:
-            self.api = f"https://api.telegram.org/bot{TELE_TOKEN}/sendMessage"
-
-    async def send(self, msg):
-        if not self.tg_enabled:
-            return
-        try:
-            await aiohttp.ClientSession().post(
-                self.api, json={"chat_id": TELE_CHAT, "text": msg}, timeout=10
-            )
-        except Exception as exc:  # noqa: BLE001
-            logging.getLogger(__name__).warning("Notifier error: %s", exc)
-
-
-NAV_G = Gauge("bot_nav_sol", "Current NAV in SOL")
-PNL_G = Gauge("bot_pnl_pct", "PnL % from start")
-
-
-# -------------------- engine ------------------------------
-class CopyEngine:
-    def __init__(self, seed_addrs: Sequence[str]):
-        self.addrs = set(seed_addrs)
-        self.notif = Notifier()
-        self.gmgn = GmgnAPI(self.notif)
-        self.sol = SolscanAPI(self.notif)
-        self.exec = JupiterExec(self.notif)
-        self.pb = PositionBook(100)
-        self.safe = SafetyChecker(self.sol, self.notif)
-        self.closed: Dict[str, float] = {}
-        self.start_nav = self.pb.init
-
-    # Kelly + ATR sizing
-    async def _size(self, token: str, sharpe: float, nav: float):
-        vol = await pyth_atr(token, notif=self.notif) or 0.05
-        edge = max(sharpe, 0)
-        f = 0.5 * edge  # half‑Kelly approximation
-        f = min(f, MAX_KELLY_F)
-        stake_pct = f / max(vol, 1e-3)
-        return max(0, min(stake_pct, 0.25)) * nav
-
-    async def _execute_buy(self, ev):
-        token = ev["token"]
-        price = float(ev["price"])
-        nav = self.pb.nav()
-        stake = await self._size(token, 1.5, nav)  # assume sharpe proxy 1.5 for now
-        amt = int(stake / price)
-        quote = await self.exec.quote(token, token, amt)  # self‑swap for placeholder
-        tx_b64 = await self.exec.swap_tx(quote["data"][0])
-
-        tx_bytes = base64.b64decode(tx_b64)
-        tx_bytes = add_priority_fee(tx_bytes)
-        kp = Keypair.from_secret_key(base58.b58decode(PRIV_KEY))
-        tx = Transaction.deserialize(tx_bytes)
-        tx.sign(kp)
-        client = Client(RPC_URL)
-        try:
-            sig = client.send_raw_transaction(tx.serialize()).get("result")
-            logging.getLogger(__name__).info("tx %s sent", sig)
-        except Exception as exc:
-            logging.getLogger(__name__).warning("tx failure: %s", exc)
-
-        await self.notif.send(f"BUY {amt} {token[:4]}… @ {price:.6f} SOL")
-        self.pb.update(token, amt, price, "buy")
-        nav = self.pb.nav()
-        NAV_G.set(nav)
-        PNL_G.set((nav - self.pb.init) / self.pb.init * 100)
-        self.pb.update_peak(nav)
-        dd = self.pb.global_dd(nav)
-        if dd >= GLOBAL_DD_PCT:
-            await self.notif.send(
-                f"GLOBAL DD {dd:.1f}% exceeds {GLOBAL_DD_PCT}% – shutting down"
-            )
-            raise SystemExit("drawdown limit hit")
-
-    async def _wallet_stream(self):
-        backoff = 1
-        while True:
-            try:
-                async with websockets.connect("wss://ws.gmgn.ai/v1") as ws:
-                    backoff = 1
-                    async for msg in ws:
-                        ev = json.loads(msg)
-                        if ev.get("address") in self.addrs and ev.get("side") == "buy":
-                            token = ev.get("token")
-                            if not token:
-                                continue
-                            if await self.safe.is_safe(token):
-                                await self._execute_buy(ev)
-            except Exception as exc:  # noqa: BLE001
-                logging.getLogger(__name__).warning(
-                    "WS reconnect in %s sec due to %s", backoff, exc
-                )
-                try:
-                    await self.notif.send(f"WS disconnect: {exc}")
-                except Exception:  # noqa: BLE001
-                    pass
-                await asyncio.sleep(backoff)
-                backoff = min(backoff * 2, 60)
-
-    async def _mark_positions(self):
-        """Update marks, close on SL/TP and prune old closed entries."""
-        while True:
-            now = time.time()
-            for token, pos in list(self.pb.pos.items()):
-                price = await pyth_price(token, notif=self.notif)
-                if not price:
-                    continue
-                self.pb.mark[token] = price
-                if price <= pos.sl or price >= pos.tp:
-                    pnl = price * pos.qty - pos.value
-                    self.pb.init += pnl
-                    self.pb.pos.pop(token, None)
-                    self.pb.mark.pop(token, None)
-                    self.closed[token] = now
-                    await self.notif.send(
-                        f"SELL {int(pos.qty)} {token[:4]}… @ {price:.6f} PnL {pnl:+.4f}"
-                    )
-
-            nav = self.pb.nav()
-            self.pb.update_peak(nav)
-            NAV_G.set(nav)
-            if self.start_nav:
-                PNL_G.set(100 * (nav - self.start_nav) / self.start_nav)
-
-            for token, ts in list(self.closed.items()):
-                if now - ts >= PRUNE_INTERVAL_H * 3600:
-                    self.closed.pop(token, None)
-
-            await asyncio.sleep(60)
-
-    async def run(self):
-        start_http_server(9100)
-        await asyncio.gather(self._wallet_stream(), self._mark_positions())
-
-
-# --------------------- main -------------------------------
-async def main():
-    seed = [w["address"] for w in gmgn().getTrendingWallets()["data"]][:10]
-    eng = CopyEngine(seed)
-    await eng.run()
-
+from gmgn_wallet_bot import main
 
 if __name__ == "__main__":
-    try:
-        asyncio.run(main())
-    except KeyboardInterrupt:
-        pass
+    main()
