+++ tests/test_safety.py
@@ -4,7 +4,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 import conftest  # noqa:F401
 import pytest
-from main import SafetyChecker
+from safety import SafetyChecker
 
 
 @pytest.mark.asyncio
