+++ tests/test_sizing.py
@@ -4,7 +4,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 import conftest  # noqa:F401
 import pytest
-from main import CopyEngine
+from engine import CopyEngine
 
 
 @pytest.mark.asyncio
@@ -14,7 +14,7 @@ async def test_size_basic(monkeypatch):
     async def atr(_, minutes=None):
         return 0.1
 
-    monkeypatch.setattr("main.pyth_atr", atr)
+    monkeypatch.setattr("sizing.pyth_atr", atr)
     size = await eng._size("TOKEN", 1.0, 100.0)
     assert size == pytest.approx(25.0)
 
@@ -26,6 +26,6 @@ async def test_size_negative_sharpe(monkeypatch):
     async def atr(_, minutes=None):
         return 0.1
 
-    monkeypatch.setattr("main.pyth_atr", atr)
+    monkeypatch.setattr("sizing.pyth_atr", atr)
     size = await eng._size("TOKEN", -1.0, 100.0)
     assert size == 0
