+++ tests/test_wallet.py
@@ -4,7 +4,7 @@ from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 import conftest  # noqa:F401
 import pytest
-from main import WalletAnalyzer, WalletMetrics
+from wallet import WalletAnalyzer, WalletMetrics
 
 
 @pytest.mark.asyncio
